Add separate folder for homework
