+++ AdvancedDataTypes/homework/homework_advanced.py
