+++ AdvancedDataTypes/homework/tests_advanced.py
@@ -1,6 +1,6 @@
 import unittest
 
-import homework_advanced
+from homework import homework_advanced
 
 
 class TestGenerateAlphabet(unittest.TestCase):
