+++ AdvancedDataTypes/lection/dict_comprehension.py
