+++ AdvancedDataTypes/lection/functional_programming.py
