+++ AdvancedDataTypes/lection/id_function.py
