+++ AdvancedDataTypes/lection/list_comprehension.py
