+++ AdvancedDataTypes/lection/type_function.py
